--- core/event-service/src/main/java/ru/practicum/event/service/EventServiceApplication.java
@@ -1,18 +1,0 @@
-package ru.practicum.event.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import ru.practicum.interaction.api.feignClient.client.request.AdminParticipationRequestClient;
-import ru.practicum.interaction.api.feignClient.client.stat.StatClient;
-import ru.practicum.interaction.api.feignClient.client.user.UserClient;
-
-@SpringBootApplication
-@EnableFeignClients(clients = {StatClient.class, UserClient.class, AdminParticipationRequestClient.class})
-public class EventServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EventServiceApplication.class, args);
-    }
-
-}
