--- core/event-service/src/main/java/ru/practicum/event/service/category/controller/AdminCategoryController.java
@@ -1,46 +1,0 @@
-package ru.practicum.event.service.category.controller;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import ru.practicum.event.service.category.service.CategoryService;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.interaction.api.dto.category.CategoryDto;
-import ru.practicum.interaction.api.dto.category.NewCategoryDto;
-import ru.practicum.interaction.api.exception.DuplicateException;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/admin/categories")
-@Slf4j
-public class AdminCategoryController {
-
-    private final CategoryService categoryService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public CategoryDto createCategory(@RequestBody @Valid NewCategoryDto newCategoryDto) {
-        log.info("Пришел запрос на создание категории.");
-        return categoryService.createCategory(newCategoryDto);
-    }
-
-    @DeleteMapping("/{catId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCategory(@PathVariable Long catId) {
-        log.info("Пришел запрос на удаление категории.");
-        categoryService.deleteCategory(catId);
-    }
-
-    @PatchMapping("/{catId}")
-    public CategoryDto updateCategory(@RequestBody @Valid CategoryDto categoryDto,
-                                      @PathVariable Long catId) {
-        log.info("Пришел запрос на обновление категории.");
-        try {
-            return categoryService.updateCategory(categoryDto, catId);
-        } catch (DataIntegrityViolationException e) {
-            throw new DuplicateException("Категория с таким именем уже существует");
-        }
-    }
-}
