--- core/event-service/src/main/java/ru/practicum/event/service/compilation/controller/AdminCompilationController.java
@@ -1,43 +1,0 @@
-package ru.practicum.event.service.compilation.controller;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.event.service.compilation.service.CompilationService;
-import ru.practicum.interaction.api.dto.compilation.CompilationDto;
-import ru.practicum.interaction.api.dto.compilation.NewCompilationDto;
-import ru.practicum.interaction.api.dto.compilation.UpdateCompilationRequest;
-
-@RestController
-@RequestMapping("/admin/compilations")
-@Slf4j
-@RequiredArgsConstructor
-public class AdminCompilationController {
-
-    private final CompilationService compilationService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public CompilationDto addCompilation(@Valid @RequestBody NewCompilationDto newCompilationDto) {
-        log.info("Поступил запрос на добавление подборки событий");
-        return compilationService.addCompilation(newCompilationDto);
-    }
-
-    @PatchMapping("/{compId}")
-    public CompilationDto updateCompilation(@Valid @RequestBody UpdateCompilationRequest updateCompilation,
-                                            @PathVariable Long compId) {
-        log.info("Поступил запрос на обновление подборки событий");
-        return compilationService.updateCompilation(compId, updateCompilation);
-    }
-
-    @DeleteMapping("/{compId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCompilation(@PathVariable Long compId) {
-        log.info("Поступил запрос на удаление подборки событий");
-        compilationService.deleteCompilation(compId);
-    }
-
-
-}
