--- core/event-service/src/main/java/ru/practicum/event/service/compilation/model/Compilation.java
@@ -1,27 +1,0 @@
-package ru.practicum.event.service.compilation.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-import ru.practicum.event.service.event.model.Event;
-
-import java.util.Set;
-
-@Entity
-@Table(name = "compilations")
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class Compilation {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "compilation_id")
-    Long id;
-    Boolean pinned;
-    String title;
-    @ManyToMany(fetch = FetchType.EAGER)
-    @JoinTable(joinColumns = @JoinColumn(name = "compilation_id"),
-            inverseJoinColumns = @JoinColumn(name = "event_id"))
-    Set<Event> events;
-}
