--- core/event-service/src/main/java/ru/practicum/event/service/compilation/repository/CompilationRepository.java
@@ -1,12 +1,0 @@
-package ru.practicum.event.service.compilation.repository;
-
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.event.service.compilation.model.Compilation;
-
-import java.util.List;
-
-public interface CompilationRepository extends JpaRepository<Compilation, Long> {
-
-    List<Compilation> findAllByPinned(Pageable pageable, Boolean pinned);
-}
