--- core/event-service/src/main/java/ru/practicum/event/service/compilation/service/CompilationService.java
@@ -1,20 +1,0 @@
-package ru.practicum.event.service.compilation.service;
-
-import ru.practicum.interaction.api.dto.compilation.CompilationDto;
-import ru.practicum.interaction.api.dto.compilation.NewCompilationDto;
-import ru.practicum.interaction.api.dto.compilation.UpdateCompilationRequest;
-
-import java.util.List;
-
-public interface CompilationService {
-
-    CompilationDto addCompilation(NewCompilationDto newCompilationDto);
-
-    CompilationDto updateCompilation(Long compId, UpdateCompilationRequest updateCompilationRequest);
-
-    void deleteCompilation(Long compId);
-
-    List<CompilationDto> getAllCompilations(Boolean pinned, Integer from, Integer size);
-
-    CompilationDto getCompilationById(Long compId);
-}
