--- core/event-service/src/main/java/ru/practicum/event/service/error/ErrorResponse.java
@@ -1,4 +1,0 @@
-package ru.practicum.event.service.error;
-
-public record ErrorResponse(String error) {
-}
