--- core/event-service/src/main/java/ru/practicum/event/service/event/repository/EventRepository.java
@@ -1,17 +1,0 @@
-package ru.practicum.event.service.event.repository;
-
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
-import ru.practicum.event.service.event.model.Event;
-
-import java.util.List;
-
-public interface EventRepository extends JpaRepository<Event, Long>, QuerydslPredicateExecutor<Event> {
-
-    List<Event> findAllByInitiatorId(Long userId, PageRequest pageRequest);
-
-    Boolean existsByCategoryId(Long catId);
-
-    List<Event> findAllByIdIn(List<Long> eventIds);
-}
