+++ core/event-service/src/main/java/ru/practicum/ewm/category/controller/PublicCategoryController.java
@@ -1,0 +1,34 @@
+package ru.practicum.ewm.category.controller;
+
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.ewm.category.service.CategoryService;
+import ru.practicum.interaction.api.dto.category.CategoryDto;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/categories")
+@Slf4j
+public class PublicCategoryController {
+
+    private final CategoryService categoryService;
+
+    @GetMapping
+    public List<CategoryDto> getCategories(
+            @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Integer from,
+            @Positive @RequestParam(name = "size", defaultValue = "10") Integer size) {
+        log.info("Пришел запрос на получение списка категорий");
+        return categoryService.getCategories(from, size);
+    }
+
+    @GetMapping("/{catId}")
+    public CategoryDto getCategoryById(@PathVariable Long catId) {
+        log.info("Пришел запрос на получение категории по id.");
+        return categoryService.getCategoryById(catId);
+    }
+}
