+++ core/event-service/src/main/java/ru/practicum/ewm/category/repository/CategoryRepository.java
@@ -1,0 +1,8 @@
+package ru.practicum.ewm.category.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.practicum.ewm.category.model.Category;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    boolean existsByName(String name);
+}
