+++ core/event-service/src/main/java/ru/practicum/ewm/compilation/repository/CompilationRepository.java
@@ -1,0 +1,12 @@
+package ru.practicum.ewm.compilation.repository;
+
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.practicum.ewm.compilation.model.Compilation;
+
+import java.util.List;
+
+public interface CompilationRepository extends JpaRepository<Compilation, Long> {
+
+    List<Compilation> findAllByPinned(Pageable pageable, Boolean pinned);
+}
