+++ core/event-service/src/main/java/ru/practicum/ewm/event/repository/EventRepository.java
@@ -1,0 +1,17 @@
+package ru.practicum.ewm.event.repository;
+
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
+import ru.practicum.ewm.event.model.Event;
+
+import java.util.List;
+
+public interface EventRepository extends JpaRepository<Event, Long>, QuerydslPredicateExecutor<Event> {
+
+    List<Event> findAllByInitiatorId(Long userId, PageRequest pageRequest);
+
+    Boolean existsByCategoryId(Long catId);
+
+    List<Event> findAllByIdIn(List<Long> eventIds);
+}
