+++ core/event-service/src/main/resources/application.yaml
@@ -18,8 +18,6 @@ eureka:
     serviceUrl:
       defaultZone: http://localhost:8761/eureka
   instance:
-    preferIpAddress: true
     hostname: localhost
     instance-id: "${spring.application.name}:${random.value}"
-    leaseRenewalIntervalInSeconds: 10
